Graph generation updates.
